Separate test files for each component
